requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
alembic==1.4.0            # via flask-migrate
certifi==2019.11.28       # via requests
chardet==3.0.4            # via requests
click==7.0                # via flask
flask-logconfig==0.4.2
flask-migrate==2.5.2
flask-negotiate==0.1.0
flask-script==2.0.6
flask-sqlalchemy==2.4.1
flask==1.1.1
gunicorn==19.9.0
idna==2.8                 # via requests
itsdangerous==1.1.0       # via flask
jinja2==2.11.1            # via flask
logconfig==0.4.0          # via flask-logconfig
logutils==0.3.5           # via logconfig
mako==1.1.1               # via alembic
markupsafe==1.1.1         # via jinja2, mako
psycopg2-binary==2.8.4
python-dateutil==2.8.1    # via alembic
python-editor==1.0.4      # via alembic
pyyaml==5.3               # via logconfig
requests==2.22.0
setuptools==43.0.0        # Latest versions Python 3.4 is unsupported
six==1.14.0               # via python-dateutil
sqlalchemy==1.3.13        # via alembic, flask-sqlalchemy
urllib3==1.25.8           # via requests
werkzeug==0.15.5          # via flask
git+http://192.168.249.38/dps-beta/common-utilities.git@1.9.0